WIP port aca-branding analyser to typescript
